Add test cases for `#at`
This change was accompanied by a few others. Notably the
`::from_values` static method allows me to more easily construct
full lists without having to pollute the test cases with a bunch
of appends or prepends or whatever. It's also just nice to have.
It'd suck to have to push values to an array or set keys in a
hash when you know what you want upfront. It makes sense for a
data structure's API to have this type of endpoint.